% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contract_df}
\alias{contract_df}
\title{Medicare Part C&D Contracts}
\format{
A data frame with 10,696 rows and 8 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{contract_id}{The health plan's Medicare contract ID}
  \item{org_type}{The type of Medicare plan}
  \item{contract_name}{Contract name}
  \item{marketing_name}{Marketing name}
  \item{parent_name}{Parent company name}
  \item{part_d_cut_point_type}{Either "Part D MA-PD" for Part C&D plans or "Part D PDP" for Part D only}
  \item{contract_type}{A categorization of contracts used to compute summary star ratings}
}
}
\usage{
contract_df
}
\description{
One record per Medicare contract ID and star year.
}
\keyword{datasets}
