% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{contracts_with_2022_summary_scores}
\alias{contracts_with_2022_summary_scores}
\title{Which contracts have enough measure star ratings for a summary star rating?}
\usage{
contracts_with_2022_summary_scores(df22)
}
\arguments{
\item{df22}{Dataframe of 2022 data, including the columns `measure_id`,
`contract_type` and `stars`}
}
\value{

}
\description{
Depending on the `contract_type`, there are different requirements for how
many measure-level star ratings are required for a health plan to receive a
summary star rating (Part C, Part D or Overall). See the 2022 Technical
Notes for details.
}
