% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{cut_measure_stars}
\alias{cut_measure_stars}
\title{Compute measure-level star ratings}
\usage{
cut_measure_stars(df, cutting_df)
}
\arguments{
\item{df}{A dataframe of measure values, including the columns `measure_id`,
`cut_point_type` and `value`}

\item{cutting_df}{A dataframe of the form output by `make_cutting_df()`}
}
\value{

}
\description{
Compute measure-level star ratings
}
