% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disenrollment_df}
\alias{disenrollment_df}
\title{Disenrollment Data}
\format{
A data frame with 11,850 rows and 5 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{contract_id}{The health plan's Medicare contract ID}
  \item{disenrollment_reason}{Reason for disenrolling from the Medicare plan}
  \item{disenrollment_pcg}{Percentage of members who disenrolled for the given reason}
  \item{disenrollment_note}{If `disenrollment_pcg` is null, this field gives an explanation}
}
}
\usage{
disenrollment_df
}
\description{
One row per contract, star year (2020 or later), and disenrollment reason.
}
\keyword{datasets}
