% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_stars_df}
\alias{domain_stars_df}
\title{Domain-Level Star Ratings}
\format{
A data frame with 89,263 rows and 6 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{domain_id}{The domain's ID code for the given year}
  \item{domain_name}{Domain name}
  \item{contract_id}{The health plan's Medicare contract ID}
  \item{stars}{Domain-level star rating}
  \item{stars_note}{If `stars` is null, this field gives an explanation}
}
}
\usage{
domain_stars_df
}
\description{
One row per contract, measure domain, and star year. This gives each
contract's domain-level star ratings.
}
\keyword{datasets}
