% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{hierarchical_clustering_cut_points}
\alias{hierarchical_clustering_cut_points}
\title{Compute cut points for a single non-CAHPS, pre-2022 measure with
Hierarchical Clustering}
\usage{
hierarchical_clustering_cut_points(value_vec, higher_is_better)
}
\arguments{
\item{value_vec}{A vector of all health plans' measure values for a given
measure and year}

\item{higher_is_better}{TRUE if higher measure values mean higher star
ratings, and FALSE otherwise}
}
\value{
A dataframe with two columns: an integer star rating in `stars`, and
a float cut-off measure value for that star rating in `cut_point`
}
\description{
Compute cut points for a single non-CAHPS, pre-2022 measure with
Hierarchical Clustering
}
