% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{make_cutting_df}
\alias{make_cutting_df}
\title{Make a dataframe of functions to convert measure values to star ratings}
\usage{
make_cutting_df(cut_point_df)
}
\arguments{
\item{cut_point_df}{Dataframe of the form output by
`mean_resampling_cut_points()`}
}
\value{

}
\description{
Make a dataframe of functions to convert measure values to star ratings
}
