% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{mean_resampling_cut_points}
\alias{mean_resampling_cut_points}
\title{Compute cut points for a single non-CAHPS measure for star year 2022}
\usage{
mean_resampling_cut_points(value_vec, higher_is_better, seed = NULL)
}
\arguments{
\item{value_vec}{A vector of all health plans' measure values for a given
measure and year}

\item{higher_is_better}{TRUE if higher measure values mean higher star
ratings, and FALSE otherwise}

\item{seed}{An optional integer passed to `set.seed()` before creating the
10 folds partitioning `value_vec`}
}
\value{
A dataframe with two columns: an integer star rating in `stars`, and
a float cut-off measure value for that star rating in `cut_point`
}
\description{
Compute cut points for a single non-CAHPS measure for star year 2022
}
