% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_df}
\alias{measure_df}
\title{Measures}
\format{
A data frame with 734 rows and 15 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{measure_id}{The ID code used by CMS to identify the measure in the
    given `star_year`}
  \item{measure_name}{The measure's name}
  \item{part}{"C" or "D"}
  \item{domain_id}{The domain's ID code for the given year}
  \item{domain_name}{Domain name}
  \item{measure_code}{All records with a given `measure_code` represent the
    same measure, over multiple years. This code was created for the R package
    and is not in the CMS data.}
  \item{measure_standard_name}{Name taken from the Star Ratings Measure
    History attachment of the Star Ratings Technical Notes PDF}
  \item{data_source}{}
  \item{data_start_date}{Start date of the period during which the measure
    data was collected}
  \item{data_end_date}{End date of the period during which the measure
    data was collected}
  \item{weight_category}{Measures in the same `weight_category` generally have the same weight}
  \item{part_weight}{The weight assigned to this measure when computing Part C/D summary star ratings}
  \item{overall_weight}{The weight assigned to this measure when computing overall star ratings}
  \item{display_measure}{TRUE if this is a display measure; otherwise, FALSE}
  \item{higher_is_better}{TRUE if higher measure values mean higher star ratings. Otherwise, FALSE}
}
}
\usage{
measure_df
}
\description{
One record per measure and star year. The table includes display measures,
which are
}
\keyword{datasets}
