% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_score_df}
\alias{measure_score_df}
\title{Measure-Level Values and Star Ratings}
\format{
A data frame with 791,087 rows and 11 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{measure_id}{The ID used by CMS to identify a measure in the given `star_year`}
  \item{measure_code}{A measure ID created for this dataset to represent one measure}
  \item{contract_id}{The health plan's Medicare contract ID}
  \item{value}{The measure value}
  \item{stars}{The star rating}
  \item{cut_point_type}{One of "Part C", "Part D MA-PD" or "Part D PDP"}
  \item{value_note}{If `value` is null, this field gives an explanation}
  \item{stars_note}{If `stars` is null, this field gives an explanation}
  \item{value_original_text}{Text from the source CSV file from CMS}
  \item{stars_original_text}{Text from the source CSV file from CMS}
}
}
\usage{
measure_score_df
}
\description{
One row per contract, measure, and star year. Each record contains the
measure value and star rating the contract received for that measure in the
given year, or a note explaining why the value/rating is NULL.
}
\keyword{datasets}
