% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_star_range_df}
\alias{measure_star_range_df}
\title{Measure-Level Star Ranges}
\format{
A data frame with 4509 rows and 8 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{cut_point_type}{One of "Part C", "Part D MA-PD" or "Part D PDP"}
  \item{measure_id}{The ID code used by CMS to identify in the given `star_year`}
  \item{stars}{Star rating}
  \item{lower_bound}{The lower bound for measure values with this star rating}
  \item{upper_bound}{The upper bound for measure values with this star rating}
  \item{higher_is_better}{TRUE if higher measure values mean higher star ratings. Otherwise, FALSE}
  \item{original_text}{Text from the source CSV file from CMS}
}
}
\usage{
measure_star_range_df
}
\description{
One row per measure, star year, and star rating. Each record gives the
measure value's lower and upper bounds. This table contains the same data
as `cut_point_df`, but rearranged.
}
\keyword{datasets}
