% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{non_cahps_2022_cut_points}
\alias{non_cahps_2022_cut_points}
\title{Compute cut points for multiple non-CAHPS 2022 measures}
\usage{
non_cahps_2022_cut_points(df, common_seed = NULL)
}
\arguments{
\item{df}{A dataframe with columns `measure_id`, `measure_code`,
`cut_point_type`, `higher_is_better` and `value`}

\item{common_seed}{A integer used as the seed for each measure's cut point
computation}
}
\value{

}
\description{
Compute cut points for multiple non-CAHPS 2022 measures
}
