% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{overall_stars_2022}
\alias{overall_stars_2022}
\title{Compute overall summary star ratings}
\usage{
overall_stars_2022(df22)
}
\arguments{
\item{df22}{A dataframe with 2022 measure data, including the columns
`measure_id`, `contract_id`, and `stars`}
}
\value{

}
\description{
Compute overall summary star ratings
}
