% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{part_cd_summary_stars_2022}
\alias{part_cd_summary_stars_2022}
\title{Compute Part C&D summary star ratings}
\usage{
part_cd_summary_stars_2022(df22)
}
\arguments{
\item{df22}{A dataframe with 2022 measure data, including the columns
`measure_id`, `contract_id`, `cut_point_type` and `stars`}
}
\value{

}
\description{
Compute Part C&D summary star ratings
}
