% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_stars_df}
\alias{summary_stars_df}
\title{Summary Star Ratings}
\format{
A data frame with 27,150 rows and 5 variables:
\describe{
  \item{star_year}{The star rating year}
  \item{summary_type}{The level of the star rating: "Part C", "Part D" or "Overall"}
  \item{contract_id}{The health plan's Medicare contract ID}
  \item{stars}{Star score}
  \item{stars_note}{If `stars` is null, this field gives an explanation}
}
}
\usage{
summary_stars_df
}
\description{
One row per contract, star year, and summary level (Part C, Part D or
Overall). This table gives contracts' highest-level star ratings.
}
\keyword{datasets}
